% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntos.R
\name{getCellOnto}
\alias{getCellOnto}
\alias{getCellLineOnto}
\alias{getEFOOnto}
\alias{getChebiLite}
\alias{getCellosaurusOnto}
\alias{getUBERON}
\alias{getChebiOnto}
\alias{getOncotreeOnto}
\alias{getDiseaseOnto}
\alias{getGeneOnto}
\alias{getHCAOnto}
\alias{getPROnto}
\alias{getPATOnto}
\alias{getCaroOnto}
\alias{getMondoOnto}
\title{load ontologies that may include non-ascii strings and therefore cannot be in data folder
useNew}
\usage{
getCellOnto()

getCellLineOnto()

getEFOOnto()

getChebiLite()

getCellosaurusOnto()

getUBERON()

getChebiOnto()

getOncotreeOnto()

getDiseaseOnto()

getGeneOnto()

getHCAOnto()

getPROnto()

getPATOnto()

getCaroOnto()

getMondoOnto()
}
\value{
instance of ontology_index (S3) from ontologyIndex

instance of ontology_index (S3) from ontologyIndex
}
\description{
load ontologies that may include non-ascii strings and therefore cannot be in data folder
useNew
}
\note{
Provenance information is kept in the form
of excerpts of top records in `dir(system.file("obo", package="ontoProc"), full=TRUE)`

getCellOnto  loads ontoRda/cellOnto.rda.rda produced from cl.owl at https://raw.githubusercontent.com/obophenotype/cell-ontology/master/cl.owl  04/24/2021,
ROBOT version 1.81 was used to convert OWL to OBO.

getCellLineOnto loads ontoRda/cellLineOnto.rda produced from clo.owl at https://raw.githubusercontent.com/CLO-ontology/CLO/master/src/ontology/clo.owl  4/22/21,
ROBOT version 1.81 was used to convert OWL to OBO.

getEFOOnto  loads ontoRda/efoOnto.rda produced from efo.owl at https://www.ebi.ac.uk/efo/efo.obo 04/15/2021,
ROBOT version 1.81 was used to convert OWL to OBO.

getChebiLite  loads ontoRda/chebi_lite.rda produced from chebi_lite.obo at http://purl.obolibrary.org/obo/chebi/chebi_lite.obo 04/01/2021,

getCellosaurusOnto loads ontoRda/cellosaurusOnto.rda produced from cellosaurus.obo at https://ftp.expasy.org/databases/cellosaurus/cellosaurus.obo  02/25/2021,

getUBERON  loads ontoRda/uberon.rda produced from uberon.owl at https://raw.githubusercontent.com/obophenotype/uberon/master/uberon.owl 04/24/2021,
ROBOT version 1.81 was used to convert OWL to OBO.

getChebiOnto  loads ontoRda/chebi_full.rda produced from chebi.obo at http://purl.obolibrary.org/obo/chebi.obo  04/01/2021,

getOncotreeOnto loads ontoRda/oncotree.rda

getDiseaseOnto  loads ontoRda/diseaseOnto.rda produced from HumanDO.obo  at https://raw.githubusercontent.com/DiseaseOntology/HumanDiseaseOntology/master/src/ontology/HumanDO.obo  04/16/2021,

getGeneOnto loads ontoRda/goOnto.rda produced from go.obo at http://purl.obolibrary.org/obo/go.obo 02/03/2021,

getHCAOnto loads ontoRda/hcaOnto.rda produced from hcao.owl at https://github.com/HumanCellAtlas/ontology/releases/tag/1.0.6 2/11/2019,
python pronto was used to convert OWL to OBO.

getPROnto loads ontoRda/PRonto.rda, produced from http://purl.obolibrary.org/obo/pr.obo 'reasoned' ontology from OBO foundry, 02-08-2019.
In contrast to other ontologies, this is imported via get_OBO with
`extract_tags='minimal'`.

getPATOnto loads ontoRda/patoOnto.rda, produced from https://raw.githubusercontent.com/pato-ontology/pato/master/pato.obo from OBO foundry, 02-08-2019.

getCaroOnto loads ontoRda/caro.rda, produced from https://raw.githubusercontent.com/obophenotype/caro/master/src/ontology/caro.owl , 04-22-2021.
}
\examples{
co = getCellOnto()
co
clo = getCellLineOnto()
length(clo$id)
che = getChebiLite()
length(che$id)
efo = getEFOOnto()
length(efo$id)
}
