% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontoInfo}
\alias{ontoInfo}
\title{A dataframe containing version information for the ontologies used in ontoProc}
\format{
data.frame instance
}
\usage{
ontoInfo
}
\description{
A dataframe containing version information for the ontologies used in ontoProc
}
\note{
A small data frame that containing the version information , date downloaded , and link to all the ontologies.
}
\examples{
head(ontoInfo) 
}
\keyword{datasets}
